+++ example/macos/Runner/AppDelegate.swift
@@ -3,12 +3,6 @@ import FlutterMacOS
 
 @main
 class AppDelegate: FlutterAppDelegate {
-  override func applicationDidFinishLaunching(_ notification: Notification) {
-    let controller = mainFlutterWindow?.contentViewController as! FlutterViewController
-    MlxLocalllmPlugin.register(with: controller.registrar(forPlugin: "MlxLocalllmPlugin"))
-    super.applicationDidFinishLaunching(notification)
-  }
-
   override func applicationShouldTerminateAfterLastWindowClosed(_ sender: NSApplication) -> Bool {
     return true
   }